requirements.txt
django==3.2.12
gunicorn
google-auth
google-cloud-pubsub
google-cloud-container
google-cloud-secret-manager 